setup: enhance --new-systemd-unit
Do not left created dropin directory in case of some initdb
failure, or so.  This allow user run the same --new-systemd-unit
repeatedly without changing the command-line.

* postgresql-setup.in (exit_handler): New exit callback wrapper.
($cleanup_dropin): Code to be executed to clean drop-in directory.
(generate_systemd_dropin): Fill $cleanup_dropin callback.
(initdb): Clean $cleanup_dropin in case of successful initdb.
This is not needed for upgrade() because we never create drop-ins
there.

* NEWS: Document changes.