postgresql-setup: reflect pg_upgrade movement
The pg_upgrade binary is not in postgresql-upgrade package now.
In some upgrade scenarios, postgresql-upgrade is redundant package
(SCLs/PGRPMSs).  Hack the code so postgresql-setup is able to
correctly advice which package should be installed to make the
"old" engine available.

* postgresql-setup.in (upgrade): Don't check for pg_upgrade binary
existence but rather for postgres binary, and warn user with
proper package name if missing.
* share/postgresql-setup/library.sh.in (parse_upgrade_setup):
White-list also 'package' constant name.
* configure.ac: Bump version.