Add postgresql-upgrade script
This should be a self-standing script which doesn't need a service file
for successful data directory upgrade.  It has similar functionality as
'postgresql-setup --upgrade' script -- but without the need to run it as
root/postgres user, and without the need to have '*.service' file
pre-configured.  It's as easy as run 'postgresql-upgrdae <DATADIR>'.

* bin/Makefile.inc: Administrivia.
* bin/postgresql-upgrade.in: The new template for the new script.
* doc/Makefile.inc: Automatically build the manual page for the new
script.